refactor(core): inject luau-evaluator port for unit testing
Move lute-specific code behind a `LuauEvaluator` port so unit tests can
swap it for a fake that throws synchronously, instead of waiting 10s for
`execFile` to kill `lute run` on `while true do end`. Public `loadConfig`
signature is unchanged; the seam is internal.
